Fix crash on attempt to remove an enemy that was already removed
The engine simply ignored such attempts,
until #134 made the engine also remove the
enemy from Battle.enemies_index, which
introduced a crash as it did not check if
the enemy is actually in enemies_index.
This introduces a nil check for getKey's result to
ensure that and restore the old behavior.