Only run post-load code once
After loading was finished, it would use global state to figure out
where to go from there. That is mostly gamestate switching, which would
take us out of the loading state, so that's fine. However, the option
which allows post-loading to directly take the user to a defined mod
does not immediately switch gamestates (as it has to load the mod!)
meaning that every frame, the game would try to load the mod over and
over until it finally escapes the loading state.

This could've been fixed with a simple `Gamestate.switch({})` to enter
an empty state (preventing the loading state from doing anything
further) but I think an explicit boolean is best for this.