Rewrite gamestate system, fix `love.graphics.reset`
The gamestate system from the HUMP library is a massive pain. It's
messy, and only existed in Kristal's codebase because it's what we
started with.

This commit finally rips it out and replaces it with... a few
variables, and a little bit of logic. Cleaning up the entire thing,
and implementing it in a much simpler way.

Due to the nature of gamestates, this was relatively easy, and also
exposed a few bugs (ex. preview scripts didn't have their `fade`
variable set anywhere BUT the update function, so if you did a `draw`
without an `update`...)

Not much should change on the user side, unless they're doing some
scary stuff. However, this commit does ANOTHER change, which may need
some care from users...

`love.graphics.reset` has ALWAYS been messed up. Sylvi didn't realize
she was overwriting a default LÖVE function when making `reset`, which
has lead to some classic bugs such as "shaders and blend modes
appearing on the error screen" and stuff like that. That specific
example has had some bandaids (resetting shaders on the error screen),
but the root was still there. I've known about the problem, but it's a
scary change to make... but enough is enough.

So, if this commit makes your game look a little different, that's
because you've actually been doing something wrong this entire time,
and the engine was broken, hiding away your error by accident...

TL;DR make sure to set fonts before trying to use them! Make sure to
set your color to the color you want, etc etc. Sometimes you forget.

Sometimes I do, as well.