Begin Utils split
This is a massive PR, and definitely a symptom of "we'll do it later",
knowing fully well that it gets worse and worse as time passes.

The amount of functions in `Utils` is insane, and I can't believe it's
entirely documented (thank you, contributors!) but it's far too much.
People understand black boxes more than Utils. Someone asks if a
function exists, someone replies with "Yep, it's in Utils!" and they
roll their eyes so far, they fall into their skull. Utils, a small
container class for various common utilities, has outgrown its purpose
and become a monster that nobody can tame. It's not a small container,
it makes up 90% of engine code. I made that statistic up on the spot,
but I bet you believed it anyway. That's how bad this is.

This commit tries to split Utils. Keyword: TRIES. But it's so big, so
massive, that it's intertwined in everything, like a tree planting its
roots. It's so deep in, you can find those roots anywhere you dig. So,
this commit has to not only split Utils, but make the Utils functions
redirect to their split versions, to not horribly break EVERY SINGLE
MOD IN EXISTENCE. If that's not bad enough, I did a quick search -- and
the pattern `(?<!\w)Utils\.` matches 810 times! And that's after I
replaced a ton of them!

Ranting aside... it's really bad. Not every function was moved out of
Utils unfortunately, and it seems like I'd be here for days if I were
to manually replace every single instance of `Utils` in the engine.

We'll probably keep the Utils module around for 0.10.0's release, and
start trying to remove it during 0.11.0's development cycle.

Either way... the point of this commit is to organize everything, while
NOT breaking every single mod. Aside from a few behavior changes, most
things should still work perfectly. If not, please report this to us.

Known behavior changes:
- `Utils.numberFromKey(t, name)` has been removed
- `Utils.keyFromNumber(t, number)` has been removed
- `Utils.hexToRgb(hex, value)` now supports hex codes which aren't just
  #RRGGBB, so `value` can now be a multiplier if you enter hex with
  alpha, for example #00FF00AA
- `Utils.lerp`'s table support (now handled by TableUtils.lerp) is more
  strict and will error if table sizes are mismatched

This commit needs SO MUCH TESTING before it's ready. So please, give it
a shot. Thank you.