Allow getEncounterText to return portrait and actor
This commit refactors a little bit of the encounter text internals.
"Encounter text" is what we call the flavor text in battles.
Unfortunately, this makes no sense, as "encounter text" SHOULD refer to
the FIRST flavor text you see when you enter a battle. Unfortunately,
I named it "encounter text" temporarily, and we never changed it.

So, to add a getter, I named it "getInitialEncounterText" which is...
questionable, but gets the job done. In the future, I might do the
following renames:
- `getInitialEncounterText` -> `getEncounterText`
- `getEncounterText` -> `getBattleText`

Either way, you can now return portraits and actors from those
functions.

`getInitialEncounterText` was added to be a simple override in the
place of `Encounter.text`, which is returned by default. This is so
you can override it and also return a portrait and an actor if you
wish.