Add lower limit for party health in BattleUI
DELTARUNE Chapter 3 introduces the "Swoon" mechanic, which is just
setting a party member's health to -999. This renders very weirdly
normally, as the health bar's width "inverts" and starts going to the
left instead, signifying you have negative health. The part that
renders weird is that there's no lower limit -- so a health of -999
will stretch the bar so far that it goes off-screen.

Chapter 3 fixes this by adding a lower limit -- it can't stretch past
the full width of the health bar.

This applies to ALL chapters in Kristal, as you cannot see this
behavior in DELTARUNE without modding or cheating in some way, and
while I love weird accuracy things like that, I'm sure people would
want to swoon party members no matter what chapter they're actually in.