Fix main menu crashing under certain conditions
A very easy crash to test is pressing Z and Enter on the same frame
while on the main menu. This'll crash the engine, as it'll immediately
enter the file select menu while `selected_mod` isn't set (as that only
gets set in update!) and the file select menu expects you to have
selected a mod by then.

The entire main menu is pretty weird, and I've only applied a bandaid
for the crash, where it tries to set `selected_mod` when you enter a
different state. But I don't really know why `selected_mod` exists in
the first place, as there's perfectly fine functions you can call
instead.