Add swooning
In the options table for `PartyBattler:hurt`, there's a new option
called `swoon`. If true, if the battler is to be downed, they'll be
swooned instead.

`battle/swooned` is a new animation which must be defined -- all of the
default party members have them set to their defeat animation, except
for Susie since she uses a unique one instead. Kris can never get
swooned in DELTARUNE, and Ralsei uses his normal defeat sprite...
Noelle can't be swooned either because she's not even IN chapter 3.

There's things like `Encounter:canSwoon(target)`,
`PartyBattler:swoon()`, and for bullets which should swoon the target,
`Bullet:shouldSwoon(damage, target, soul)`. Note that due to the nature
of DELTARUNE's target system, `target` is highly likely to just be the
string "ANY" in `shouldSwoon`. Thankfully, `Encounter:canSwoon` at
least gives you a battler.

`Encounter:isAutoHealingEnabled(battler)` is a new function for this,
as well. It returns true by default, but you can override it for your
encounter if you want auto healing to be disabled, for example if you
have a boss which swoons the party.

For the same effect as the Roaring Knight in chapter 3, you must:
- Make all of the bullets swoon (return `true` in `Bullet:shouldSwoon`)
- Make Kris unable to be swooned (check `target` in
  `Encounter:canSwoon`)
- Disable auto healing for the encounter (return false in
  `Encounter:isAutoHealingEnabled`)

It's very strange that auto healing being disabled is completely
independant from swooning, but we aren't here to change DELTARUNE
behavior, only implement it.