Remove "no map, shop or encounter" kick to menu
This was added in a PR to "fix a potential softlock", but if your mod
does this, it's most likely on purpose... maybe to have a custom
gameplay state or something.

The game should NOT kick you to the menu if you aren't in any of the
engine states. That's a very weird thing to assume.

The engine should not try to fix errors that modders may make.