Battle intro timer is now 13 (from 15)
A long-running accuracy problem in Kristal that has always annoyed me
was the length of the intro state being set to 15, despite it being 13
in DELTARUNE. This was because Susie's intro animation did not properly
(improperly) wrap to the first frame.

I have now realized that this was finally solved today, as every
animation in Kristal was unintentionally slower than what you'd expect,
due to the implementation of animation speeds, until I fixed that in
15db64503bb4575066a3eef41591ff73f8ed1aeb.

What this means is, we can finally drop the animation timer to 13
frames, from 15 frames, since the animation speeds are now working
properly.