Fix bad shake logic
For some reason, the object would attempt to move every frame rather
than on the delay interval. As barely anyone uses shake delays, this
was, as far as I'm aware, never reported. I've known about this for a
bit, but I didn't want to make a breaking change -- that's a little
dumb, so this has been fixed now.