Make the dark transition particles accurate
This commit changes some incorrect conversion code to the proper code,
using DELTARUNE as a reference.

Before, we were using the actor's width and height, which is NOT what
DELTARUNE does (not having an actor system). The actor's width and
height was nowhere near the width/height that DELTARUNE uses, since it
grabs the size of a specific animation. The code no longer uses the
actor's width and height, but queries the texture for its width and
height directly, fixing the problem.

This also removes an accidental debug print.