Remove testmod
Long overdue, it seems like there's nothing I can do about people
trying to use the testmod if it's still there, ignoring all of the
warnings.

The test mod is not -- and never should've been used as -- a public
resource to learn from. A place to take code from. "But it might have
some good stuff in it!" was never an excuse. I told people not to use
code from it, and they did.

I really wish that I didn't have to do this, because it sucks for the
Kristal devs & anyone who makes PRs, but we'll have to figure something
else out to only share it privately with people who know not to steal
stuff from it.

We have instead created a private repository for the test mod. If
engine contributors need (or want) access, send a message to Nyako on
Discord. This does make it more annoying, but it is what it is.